+++ pkg/image/context.go
@@ -4,9 +4,9 @@ import (
 	"io"
 )
 
-type Helm interface {
-	AddRepo(chart, repository string) error
-	Pull(chart, repository, version, destDir string) (string, error)
+type HelmClient interface {
+	AddRepo(repository *HelmRepository) error
+	Pull(chart string, repository *HelmRepository, version, destDir string) (string, error)
 	Template(chart, repository, version, valuesFilePath, kubeVersion string) ([]map[string]any, error)
 }
 
@@ -58,5 +58,5 @@ type Context struct {
 	// RPMResolver responsible for resolving rpm/package dependencies
 	RPMResolver    rpmResolver
 	RPMRepoCreator rpmRepoCreator
-	Helm           Helm
+	HelmClient     HelmClient
 }