+++ pkg/registry/helm.go
@@ -14,12 +14,18 @@ type HelmChart struct {
 	ContainerImages []string
 }
 
-func HelmCharts(helmCharts []image.HelmChart, valuesDir, buildDir, kubeVersion string, helm image.Helm) ([]*HelmChart, error) {
+func HelmCharts(helm *image.Helm, valuesDir, buildDir, kubeVersion string, helmClient image.HelmClient) ([]*HelmChart, error) {
 	var charts []*HelmChart
+	chartRepoMap := mapChartRepos(helm)
 
-	for _, helmChart := range helmCharts {
+	for _, helmChart := range helm.Charts {
 		c := helmChart
-		chart, err := handleChart(&c, valuesDir, buildDir, kubeVersion, helm)
+		r, ok := chartRepoMap[c.RepositoryName]
+		if !ok {
+			return nil, fmt.Errorf("repository not found for chart %s", c.Name)
+		}
+
+		chart, err := handleChart(&c, r, valuesDir, buildDir, kubeVersion, helmClient)
 		if err != nil {
 			return nil, fmt.Errorf("handling chart resource: %w", err)
 		}
@@ -30,7 +36,7 @@ func HelmCharts(helmCharts []image.HelmChart, valuesDir, buildDir, kubeVersion s
 	return charts, nil
 }
 
-func handleChart(chart *image.HelmChart, valuesDir, buildDir, kubeVersion string, helm image.Helm) (*HelmChart, error) {
+func handleChart(chart *image.HelmChart, repo *image.HelmRepository, valuesDir, buildDir, kubeVersion string, helmClient image.HelmClient) (*HelmChart, error) {
 	var valuesPath string
 	var valuesContent []byte
 	if chart.ValuesFile != "" {
@@ -42,12 +48,12 @@ func handleChart(chart *image.HelmChart, valuesDir, buildDir, kubeVersion string
 		}
 	}
 
-	chartPath, err := downloadChart(chart, helm, buildDir)
+	chartPath, err := downloadChart(chart, repo, helmClient, buildDir)
 	if err != nil {
 		return nil, fmt.Errorf("downloading chart: %w", err)
 	}
 
-	images, err := getChartContainerImages(chart, helm, chartPath, valuesPath, kubeVersion)
+	images, err := getChartContainerImages(chart, helmClient, chartPath, valuesPath, kubeVersion)
 	if err != nil {
 		return nil, fmt.Errorf("getting chart container images: %w", err)
 	}
@@ -65,12 +71,12 @@ func handleChart(chart *image.HelmChart, valuesDir, buildDir, kubeVersion string
 	return &helmChart, nil
 }
 
-func downloadChart(chart *image.HelmChart, helm image.Helm, destDir string) (string, error) {
-	if err := helm.AddRepo(chart.Name, chart.Repo); err != nil {
+func downloadChart(chart *image.HelmChart, repo *image.HelmRepository, helmClient image.HelmClient, destDir string) (string, error) {
+	if err := helmClient.AddRepo(repo); err != nil {
 		return "", fmt.Errorf("adding repo: %w", err)
 	}
 
-	chartPath, err := helm.Pull(chart.Name, chart.Repo, chart.Version, destDir)
+	chartPath, err := helmClient.Pull(chart.Name, repo, chart.Version, destDir)
 	if err != nil {
 		return "", fmt.Errorf("pulling chart: %w", err)
 	}
@@ -87,8 +93,8 @@ func getChartContent(chartPath string) (string, error) {
 	return base64.StdEncoding.EncodeToString(data), nil
 }
 
-func getChartContainerImages(chart *image.HelmChart, helm image.Helm, chartPath, valuesPath, kubeVersion string) ([]string, error) {
-	chartResources, err := helm.Template(chart.Name, chartPath, chart.Version, valuesPath, kubeVersion)
+func getChartContainerImages(chart *image.HelmChart, helmClient image.HelmClient, chartPath, valuesPath, kubeVersion string) ([]string, error) {
+	chartResources, err := helmClient.Template(chart.Name, chartPath, chart.Version, valuesPath, kubeVersion)
 	if err != nil {
 		return nil, fmt.Errorf("templating chart: %w", err)
 	}
@@ -104,4 +110,19 @@ func getChartContainerImages(chart *image.HelmChart, helm image.Helm, chartPath,
 	}
 
 	return images, nil
+}
+
+func mapChartRepos(helm *image.Helm) map[string]*image.HelmRepository {
+	chartRepoMap := make(map[string]*image.HelmRepository)
+
+	for _, chart := range helm.Charts {
+		for _, repo := range helm.Repositories {
+			if chart.RepositoryName == repo.Name {
+				r := repo
+				chartRepoMap[chart.RepositoryName] = &r
+			}
+		}
+	}
+
+	return chartRepoMap
 }